Remove unnecessarily public 'devpiserver_makepypistage' fixture. Refs #780
